commands: Move selected change display
Selected-change rendering belongs to command selection helpers while hunk
tracking owns selected-state navigation and persistence.

The renderer still lived in the data hunk tracker, which made command flow
depend on a lower-layer presentation shortcut.

This commit adds a command-layer selected-change display helper, points
production callers at it, and removes the renderer from hunk tracking.

Text hunk display also reads persisted line state before falling back to
the raw patch, so filtered selections use the same cached view as line
actions.
