+++ src/git_stage_batch/commands/session/iteration.py
@@ -3,10 +3,11 @@
 from __future__ import annotations
 
 from ...data.file_tracking import auto_add_untracked_files
-from ...data.hunk_tracking import fetch_next_change, show_selected_change
+from ...data.hunk_tracking import fetch_next_change
 from ...data.session import clear_iteration_state
 from ...exceptions import NoMoreHunks
 from ...i18n import _
+from ..selection.selected_change_display import show_selected_change
 
 
 def restart_iteration_pass(*, quiet: bool = False) -> None:
